March 31, 2022
- Updates to some code from week 2 assignment
- Code for loop statements are added, the questions were practiced in class

+++ week2/SumOfFirstAndFourth.java
@@ -18,32 +18,34 @@ public class SumOfFirstAndFourth {
      * 
      * 
      */
-    long number;
+    int number;
     SumOfFirstAndFourth() {
         this.number = 11111;
     }
-    SumOfFirstAndFourth(long number) {
+    SumOfFirstAndFourth(int number) {
         this.number = number;
     }
 
     // returns the sum of the first and the second last digit of the five digit number
     public long sumOfFirstAndFourth() {
         // get the first digit by dividing the number by 10
-        long firstDigit = (this.number/10000);
+        int firstDigit = (this.number/10000);
         // get the second last digit by getting the modulus of the number by 100 and then dividing the modulus by 10 and converting it to long
-        long fourthDigit =  (long)(((this.number%100) / 10));
+        int fourthDigit =  ((this.number%100) / 10);
         return (firstDigit + fourthDigit);
     }
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
-        long scannedNumber;
+        int scannedNumber;
         // asks the user to enter a five digit number until the user enters a five digit number
         do {
             System.out.println("Enter a five digit number: ");
-            scannedNumber = scan.nextLong();
+            scannedNumber = scan.nextInt();
         } while(scannedNumber < 10000 || scannedNumber > 99999);
         SumOfFirstAndFourth number = new SumOfFirstAndFourth(scannedNumber);
-        System.out.println("Sum of first and fourth digits of "+number.number+" is: "+number.sumOfFirstAndFourth());
+        System.out.println("Sum of first and fourth digits of "+
+                           number.number+" is: "+
+                           number.sumOfFirstAndFourth());
         scan.close();
     }
 }
